+++ packages/w3p/src/helpers/eth.ts
@@ -1,3 +1,4 @@
+import { DECIMALS } from '@distributedlab/tools'
 import { type providers, utils } from 'ethers'
 
 import { EIP1193, EIP1474 } from '@/enums'
@@ -26,15 +27,23 @@ export const requestSwitchEthChain = async (
 
 export const requestAddEthChain = async (
   provider: providers.Web3Provider,
-  chainId: number,
-  chainName: string,
-  chainRpcUrl: string,
+  chain: Chain,
 ): Promise<void> => {
   await provider.send('wallet_addEthereumChain', [
     {
-      chainId: utils.hexValue(chainId),
-      chainName,
-      rpcUrls: [chainRpcUrl],
+      chainId: utils.hexValue(Number(chain.id)),
+      chainName: chain.name,
+      ...(chain.token.name &&
+        chain.token.symbol && {
+          nativeCurrency: {
+            name: chain.token.name,
+            symbol: chain.token.symbol,
+            decimals: chain.token.decimals ?? DECIMALS.WEI,
+          },
+        }),
+      rpcUrls: [chain.rpcUrl],
+      blockExplorerUrls: [...(chain.explorerUrl ? [chain.explorerUrl] : [])],
+      ...(chain.icon && { iconUrls: [chain.icon] }),
     },
   ])
 }
@@ -43,45 +52,76 @@ export const connectEthAccounts = async (provider: providers.Web3Provider) => {
   await provider.send('eth_requestAccounts', [])
 }
 
-export function handleEthError(error: EthProviderRpcError): void {
+export function handleEthError(
+  error: EthProviderRpcError,
+  defaultHandlerFn?: (error: EthProviderRpcError) => unknown,
+): never {
   switch (error.code) {
     case EIP1193.UserRejectedRequest:
-      throw new errors.ProviderUserRejectedRequest()
+      throw new errors.ProviderUserRejectedRequest(error)
     case EIP1193.UnrecognizedChain:
-      throw new errors.ProviderChainNotFoundError()
+      throw new errors.ProviderChainNotFoundError(error)
     case EIP1193.Unauthorized:
-      throw new errors.ProviderUnauthorized()
+      throw new errors.ProviderUnauthorized(error)
     case EIP1193.UnsupportedMethod:
-      throw new errors.ProviderUnsupportedMethod()
+      throw new errors.ProviderUnsupportedMethod(error)
     case EIP1193.Disconnected:
-      throw new errors.ProviderDisconnected()
+      throw new errors.ProviderDisconnected(error)
     case EIP1193.ChainDisconnected:
-      throw new errors.ProviderChainDisconnected()
+      throw new errors.ProviderChainDisconnected(error)
     case EIP1474.ParseError:
-      throw new errors.ProviderParseError()
+      throw new errors.ProviderParseError(error)
     case EIP1474.InvalidRequest:
-      throw new errors.ProviderInvalidRequest()
+      throw new errors.ProviderInvalidRequest(error)
     case EIP1474.MethodNotFound:
-      throw new errors.ProviderMethodNotFound()
+      throw new errors.ProviderMethodNotFound(error)
     case EIP1474.InvalidParams:
-      throw new errors.ProviderInvalidParams()
+      throw new errors.ProviderInvalidParams(error)
     case EIP1474.InternalError:
-      throw new errors.ProviderInternalError()
+      throw new errors.ProviderInternalError(error)
     case EIP1474.InvalidInput:
-      throw new errors.ProviderInvalidInput()
+      throw new errors.ProviderInvalidInput(error)
     case EIP1474.ResourceNotFound:
-      throw new errors.ProviderResourceNotFound()
+      throw new errors.ProviderResourceNotFound(error)
     case EIP1474.ResourceUnavailable:
-      throw new errors.ProviderResourceUnavailable()
+      throw new errors.ProviderResourceUnavailable(error)
     case EIP1474.TransactionRejected:
-      throw new errors.ProviderTransactionRejected()
+      throw new errors.ProviderTransactionRejected(error)
     case EIP1474.MethodNotSupported:
-      throw new errors.ProviderMethodNotSupported()
+      throw new errors.ProviderMethodNotSupported(error)
     case EIP1474.LimitExceeded:
-      throw new errors.ProviderLimitExceeded()
+      throw new errors.ProviderLimitExceeded(error)
     case EIP1474.JsonRpcVersionNotSupported:
-      throw new errors.ProviderJsonRpcVersionNotSupported()
+      throw new errors.ProviderJsonRpcVersionNotSupported(error)
     default:
+      if (defaultHandlerFn) {
+        defaultHandlerFn(error)
+      }
+
       throw error
   }
+}
+
+export const wrapExternalEthProvider = (
+  provider: providers.ExternalProvider,
+  errorHandler?: (error: Error) => unknown,
+) => {
+  const _baseRequest = provider.request?.bind(provider)
+
+  provider.request = async (request: {
+    method: string
+    params?: Array<unknown>
+  }) => {
+    let result: unknown
+
+    try {
+      result = await _baseRequest?.(request)
+    } catch (error) {
+      handleEthError(error as EthProviderRpcError, errorHandler)
+    }
+
+    return result
+  }
+
+  return provider
 }